dashboard-backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Django==2.2
django-channels==0.7.0
django-cors-headers==3.14.0
django-decouple==2.1
django-environ==0.10.0
djangorestframework==3.11.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
